change double to float because double not supported

+++ src/Structure/AlderScheme.php
@@ -19,7 +19,7 @@ class AlderScheme
         'file-multiple'   => 'string',
         'image'           => 'string',
         'month'           => 'string',
-        'number'          => 'double',
+        'number'          => 'float',
         'password'        => 'string',
         'repeater'        => 'text',
         'select'          => 'string',
